+++ src/Common/CSharp/Analysis/DiagnosticPropertyKeys.cs
@@ -4,7 +4,6 @@ namespace Roslynator.CSharp.Analysis;
 
 internal static class DiagnosticPropertyKeys
 {
-    internal static readonly string UseCollectionExpression = nameof(UseCollectionExpression);
     internal static readonly string ImplicitToCollectionExpression = nameof(ImplicitToCollectionExpression);
     internal static readonly string CollectionExpressionToImplicit = nameof(CollectionExpressionToImplicit);
     internal static readonly string ExplicitToCollectionExpression = nameof(ExplicitToCollectionExpression);